Fix calculation for traits with same value
Traits that have exactly the same value for different species would get a value of NA in the results because we would divide by zero in calculating the distance matrix. Therefore we remove these traits before calculation. But in the end we need to add +1 for each species combination representing this specific trait because 1 means completely equal.